Do not use absolute-form for SSL requests over a proxy.
An SSL request over a proxy is sent to the origin server via
a connection established by CONNECT method to the proxy.  The
request itself is directly sent to the origin server (without
modification by the proxy).

In this case, we MUST not use absolute-form of the URI in the
request line.  This is indicated in the section 5.3.1 of RFC 7230.